+++ src/config.rs
@@ -771,7 +771,7 @@ impl<'a> de::Deserialize<'a> for ActionWrapper {
             fn expecting(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
                 f.write_str("Paste, Copy, PasteSelection, IncreaseFontSize, DecreaseFontSize, \
                             ResetFontSize, ScrollPageUp, ScrollPageDown, ScrollToTop, \
-                            ScrollToBottom, ClearHistory, Hide, ClearLogNotice, NewInstanceSameDir or Quit")
+                            ScrollToBottom, ClearHistory, Hide, ClearLogNotice, SpawnNewInstance or Quit")
             }
 
             fn visit_str<E>(self, value: &str) -> ::std::result::Result<ActionWrapper, E>
@@ -792,7 +792,7 @@ impl<'a> de::Deserialize<'a> for ActionWrapper {
                     "Hide" => Action::Hide,
                     "Quit" => Action::Quit,
                     "ClearLogNotice" => Action::ClearLogNotice,
-                    "NewInstanceSameDir" => Action::NewInstanceSameDir,
+                    "SpawnNewInstance" => Action::SpawnNewInstance,
                     _ => return Err(E::invalid_value(Unexpected::Str(value), &self)),
                 }))
             }